- Merge branch develop into fix/timestep-as-int-bug
- Merge branch develop into fix/timestep-as-int-bug
- fix: cast timestep to int

+++ mava/advanced_usage/ff_ippo_store_experience.py
@@ -618,7 +618,7 @@ def run_experiment(_config: DictConfig) -> None:  # noqa: CCR001
 
         # Log the results of the training.
         elapsed_time = time.time() - start_time
-        t = steps_per_rollout * (eval_step + 1)
+        t = int(steps_per_rollout * (eval_step + 1))
         learner_output.episode_metrics["steps_per_second"] = steps_per_rollout / elapsed_time
         learner_output.episode_metrics["timestep"] = t
 
@@ -677,7 +677,7 @@ def run_experiment(_config: DictConfig) -> None:  # noqa: CCR001
 
         elapsed_time = time.time() - start_time
 
-        t = steps_per_rollout * (eval_step + 1)
+        t = int(steps_per_rollout * (eval_step + 1))
         evaluator_output.episode_metrics["steps_per_second"] = steps_per_rollout / elapsed_time
         logger.log(evaluator_output.episode_metrics, t, eval_step, LogEvent.ABSOLUTE)
